+++ test/dummy/test/test_helper.rb
@@ -3,6 +3,11 @@
 ENV["RAILS_ENV"] ||= "test"
 ENV.delete("OPENAI_API_KEY")
 ENV.delete("GEMINI_API_KEY")
+ENV.delete("GOOGLE_AI_STUDIO")
+ENV.delete("google_ai_studio")
+ENV.delete("TYPESAFE_API_KEY")
+ENV.delete("TYPESAFE")
+ENV.delete("typesafe")
 
 require_relative "../config/environment"
 require "rails/test_help"