+++ test/test_helper.rb
@@ -4,6 +4,11 @@ $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 
 ENV.delete("OPENAI_API_KEY")
 ENV.delete("GEMINI_API_KEY")
+ENV.delete("GOOGLE_AI_STUDIO")
+ENV.delete("google_ai_studio")
+ENV.delete("TYPESAFE_API_KEY")
+ENV.delete("TYPESAFE")
+ENV.delete("typesafe")
 
 require_relative "simplecov_helper"
 require "minitest/autorun"